MNT: Large refactor splitting up contents into submodules
